tests/test_oauth_routes.py
import pytest

class TestOAuthRoutes:
    """
    Class for testing OAuth routes.

    This class contains test cases to ensure the proper functioning of OAuth login
    and callback endpoints. Each test examines a specific scenario to validate the
    behavior and responses of the OAuth-related workflows. This includes testing
    the start of the OAuth login process, handling of invalid states in the
    callback, and scenarios where the token is missing in the process.

    Attributes:
        None
    """
    def test_oauth_login_start(self, client):
        """
        Tests the initiation of the OAuth login process.

        This function tests if the OAuth login flow starts correctly by sending
        a GET request to the designated OAuth login endpoint. It verifies the
        response status code and checks the session for the presence of the
        OAuth state.

        Parameters:
        client
            Flask test client used to simulate HTTP requests.

        Returns:
        None
        """
        response = client.get('/login/oauth')
        assert response.status_code == 302
        with client.session_transaction() as sess:
            assert 'oauth_state' in sess

    def test_oauth_callback_invalid_state(self, client):
        """
        Tests the handling of an OAuth callback with an invalid state.

        This function simulates an HTTP GET request to the OAuth callback endpoint
        with an invalid state parameter and asserts that the server returns the
        expected HTTP 400 Bad Request status code.

        Args:
            client: A test client instance used to simulate HTTP requests.

        Returns:
            None
        """
        response = client.get('/authorized?state=invalid')
        assert response.status_code == 400

    def test_oauth_callback_missing_token(self, client):
        """
        Tests the scenario where the OAuth callback is invoked without a token and
        ensures that the response status code is 400, as expected.

        Args:
            client: The test client used to simulate requests in the application.

        Returns:
            None
        """
        with client.session_transaction() as session:
            session['oauth_state'] = 'test_state'
        response = client.get('/authorized?state=test_state')
        assert response.status_code == 400