requirements.txt
async-timeout==5.0.1
Authlib==1.5.2
blinker==1.9.0
certifi==2025.4.26
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
crypto==1.4.1
cryptography==44.0.2
Deprecated==1.2.18
fernet==1.0.1
Flask==3.1.0
Flask-Limiter==3.12
Flask-Login==0.6.3
Flask-Mail==0.10.0
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.2
gevent==25.4.2
greenlet==3.2.1
gunicorn==23.0.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
limits==5.1.0
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
Naked==0.1.32
ordered-set==4.1.0
packaging==25.0
pillow==11.2.1
portalocker==3.1.1
pyaes==1.6.1
pycparser==2.22
pycryptodome==3.23.0
Pygments==2.19.1
python-dotenv==1.1.0
python-dateutil==2.9.0.post0
PyYAML==6.0.2
qrcode==8.1
redis==6.0.0
requests==2.32.3
rich==13.9.4
shellescape==3.8.1
SQLAlchemy==2.0.40
typing_extensions==4.13.2
urllib3==2.4.0
Werkzeug==3.1.3
wrapt==1.17.2
WTForms==3.2.1
zope.event==5.0
zope.interface==7.2